Add files via upload
This project is still in progress and I will be continuously updating the edits.

The purpose of this project is to find correlations and dependencies between different attributes for Weather conditions & external parameters that influence Cab Booking in. Boston(MA). The main aim is to predict surge and price for the bookings analyzing the cab booking data from UBER & LYFT during the span of two months.